UI Objectives are complete
To create UI Objectives

The UI will now show objectives and Listeners when the player runs into a trigger that has a reference to those objective's listener.
When the objectives are completed they will be struck through in text.
When all objectives are completed the all Objectives and Listeners displayed in the UI will destroy

1. Place a UI Objective Trigger in the level where you want the player to be notified about objectives
2. Have an Objectives Listener in the scene with it's Objectives list populated
3. In the Listener and in the Listener's Objectives write your custom UI Message
- Listener: Big Message at top
- Objectives: Little Messages on the bottom
4. Make sure the level has a UICanvas on it